Use unsigned long ints for colors/attrs.
git-svn-id: file:///Users/kling/src/protoman/svn/trunk/ytalk@681 8120b729-f3fe-0310-9465-b9035d29392c

+++ src/main.c
@@ -28,10 +28,10 @@ char *vhost = NULL;		/* specified virtual host */
 ylong myuid;			/* global uid */
 char *gshell = NULL;		/* global shell */
 #ifdef YTALK_COLOR
-int ui_colors = 40;		/* TODO: change default */
-int ui_attr = 0;		/* UI output attributes */
-int menu_colors = 0;		/* TODO: change default */
-int menu_attr = 0;
+unsigned long int ui_colors = 40;		/* TODO: change default */
+unsigned long int ui_attr = 0;		/* UI output attributes */
+unsigned long int menu_colors = 0;		/* TODO: change default */
+unsigned long int menu_attr = 0;
 #endif
 char *title_format = NULL;
 char *user_format = NULL;
